implement(UseMob): Insert and destroy item to use while interacting with Mob.
